// types/types.ts
export type Job = {
  id: string;
  title: string;
  company: string;
  salary: string;
  saved: boolean;
  applied: boolean; 
  companyLogo: string;
  location: string;
  jobType: string;
  workModel: string;
  description: string;
  seniorityLevel: string;
  publishDate: string;
  expiryDate: string;
};